Jesse's changes to swift subsystem

+++ gisdemo/swift.py
@@ -15,7 +15,7 @@ def alarm_handler(signum, frame):
     raise Alarm
 
 
-def do_swift_command(swift_proxy, operation, bucket, *args):
+def do_swift_command(swift_proxy, operation, bucket, timeout, *args):
   command = "swift -A "+swift_proxy+" -U "+settings.SWIFT_USER+ \
     " -K "+settings.SWIFT_PWD+ " " + \
     operation + " " + \
@@ -24,20 +24,25 @@ def do_swift_command(swift_proxy, operation, bucket, *args):
 
   print command
 
+  # spawn a shell that executes swift, we set the sid of the shell so
+  #  we can kill it and all its children with os.killpg
   p = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE,
     stderr=subprocess.PIPE, preexec_fn=os.setsid)
 
-  try:
-    signal.signal(signal.SIGALRM, alarm_handler)
-    signal.alarm(3*60)  # we will timeout after 3 minutes 
+  if timeout:
+    try:
+      signal.signal(signal.SIGALRM, alarm_handler)
+      signal.alarm(3*60)  # we will timeout after 3 minutes 
 
-    p.wait()
-    signal.alarm(0)  # reset the alarm
-  except Alarm:
-    os.killpg(p.pid, signal.SIGTERM)
-    # raise an assertion so we can continue execution after 
-    #  (should really have our own exception but fk it)
-    raise AssertionError("Timeout gettimg images from swift")
+      p.wait()
+      signal.alarm(0)  # reset the alarm
+    except Alarm:
+      os.killpg(p.pid, signal.SIGTERM)
+      # raise an assertion so we can continue execution after 
+      #  (should really have our own exception but fk it)
+      raise AssertionError("Timeout "+operation+"ing swift images")
+  else:
+      p.wait()
 
   return p
 
@@ -54,10 +59,10 @@ def swift(operation, bucket, *args):
   (options, args_not_used) = parser.parse_args()
   print "Using swift proxy http://"+options.swift_host+":8080/auth/v1.0"
 
-  process = do_swift_command("http://"+options.swift_host+":8080/auth/v1.0", operation, bucket, *args)
+  process = do_swift_command("http://"+options.swift_host+":8080/auth/v1.0", operation, bucket, True, *args)
   if process.returncode != 0 and options.swift_host != DEFAULT_SWIFT_HOST:
     print "Failed on swift host %s with %s, trying on %s" % (options.swift_host, process.communicate()[1], DEFAULT_SWIFT_HOST)
-    process = do_swift_command(settings.SWIFT_PROXY, operation, bucket, *args)
+    process = do_swift_command(settings.SWIFT_PROXY, operation, bucket, True, *args)
   assert process.returncode == 0, "Failed on swift host %s with %s" % (DEFAULT_SWIFT_HOST, process.communicate()[1])
